Positioning fixes
Positioning was reeaaallly screwy, so this is an attempt to make things more consistent.

Everything now has it's pos in the bottom left corner,
except for players which are taller images and just have (0, 70),
and items, which have their position just to the right of the bottom middle of the image so they line up on conveyors properly.