Fixed the UI for placing machines down
Removed the previous dodgy calculations, and went with a simpler version - just use the grid squares to either left or right. It also displays an arrow when choosing direction, just to make it super clear.

Also fixed a bug where you might not have had enough conveyors to get all the way across to the loading dock.