More playtesting and polish
Fixed some of the machine representations, trying to avoid situations where the machine has a part on top, but also a conveyor belt hopper+item icon.

Added a timer - speed through the level and overall game time is now recorded.

There are now icons on the chutes and loading docks, and a number on the dock to show how much has been loaded.